Checking root for existence when destroy_root_taxon is called. If this is the top Taxonomy then root would be nil.

+++ core/app/models/spree/taxonomy.rb
@@ -18,7 +18,7 @@ module Spree
       end
 
       def destroy_root_taxon
-        self.root.destroy
+        self.root.destroy if self.root
       end
   end
 end
